+++ impl/mem_read.c
@@ -1,7 +1,36 @@
 #include <stdio.h>
 #include <stdint.h>
 
-#include <adaptor.h>
+#define NETTLP 1
+#define DEVMEM 2
+#define PMEM 3
+
+#include <libtlp.h>
+#include "util.h"
+
+int read_pmem(uintptr_t addr, int method)
+{
+    printf("read at 0x%lx\n", addr);
+
+    switch (method)
+    {
+    case NETTLP:
+        /* code */
+        break;
+    case DEVMEM:
+        /* code */
+        break;
+    case PMEM:
+        /* code */
+        break;
+
+    default:
+        printf("Not implemented this method\n");
+        break;
+    }
+
+    return 0;
+}
 
 int main(int argc, char **argv)
 {
@@ -13,7 +42,8 @@ int main(int argc, char **argv)
     addr = 31;
 
     printf("%d\n", DEVMEM);
-    // result = read_pmem(addr, DEVMEM);
+    result = read_pmem_testtest(1);
+    result = read_pmem(addr, DEVMEM);
 
     return 0;
 }
