--- include/devmem.h
@@ -1,4 +1,0 @@
-#ifndef _DEVMEM_H_
-#define _DEVMEM_H_
-
-#endif
