--- lib/devmem.c
@@ -1,2 +1,0 @@
-#include <adaptor.h>
-#include <devmem.h>
